+++ .bumpversion.cfg
@@ -19,7 +19,7 @@ replace = /v{new_version}...main
 search = version = release = "{current_version}"
 replace = version = release = "{new_version}"
 
-[bumpversion:file:src/idmlaser/__init__.py]
+[bumpversion:file:src/laser_core/__init__.py]
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
