+++ .readthedocs.yml
@@ -25,4 +25,4 @@ python:
 search:
   ranking:
     installation.html: 3
-    idmlaser.html: 2
+    laser_core.html: 2
