+++ docs/installation.rst
@@ -4,4 +4,4 @@ Installation
 
 At the command line::
 
-    pip install idmlaser
+    pip install laser-core
