--- docs/reference/idmlaser.rst
@@ -1,9 +1,0 @@
-idmlaser
-========
-
-.. testsetup::
-
-    from idmlaser import *
-
-.. automodule:: idmlaser
-    :members:
