+++ docs/reference/laser_core.rst
@@ -1,0 +1,9 @@
+laser-core
+==========
+
+.. testsetup::
+
+    from laser_core import *
+
+.. automodule:: laser_core
+    :members:
