+++ pytest.ini
@@ -23,6 +23,6 @@ testpaths =
 filterwarnings =
     error
 # You can add exclusions, some examples:
-#    ignore:'idmlaser' defines default_app_config:PendingDeprecationWarning::
+#    ignore:'laser_core' defines default_app_config:PendingDeprecationWarning::
 #    ignore:The {{% if:::
 #    ignore:Coverage disabled via --no-cov switch!
